Update README.md
Translatoray

There are a lot of translation apps that didn't have the other dialect languages in the Philippines.

Then I thought, why not? Translatoray Mobile App.

The purpose of this Translatoray Mobile Application is to help the people that are still struggling in
tagalog who wants to learn english. Then this app can translate the other dialect that didn't 
have in any translation apps in the Philippines.

Timeline and Workflow

Week 1-2 Planing and Research
- Conduct research and analysis about translation apps and available dialect databases.
- Collaborate with educators, linguists, and local experts to ensure accurate translations.
- Define project goals, scope, and target audience.
  
Week 3-4
- Delayed due to other subjects, activities and projects that i need to complete.
- Use spare time to brainstorm ideas or refine goals, even if progress is limited.
  
Week-5 Preparation
- Start downloading the resources needed to build my application
- Set up a basic development environment.

Week-6 Design
- Create UI/UX designs

Week-7 Initial Development 
- Begin coding the application 

Week-8
- Still on the process of making my application