Update README.md
# Eco:Explorer

Timeline and Workflow

Week 1-2 Planing and Research
- Conduct research and analysis about Eco:explorer apps and available existing app in the appstore.
- Collaborate with the graduates in IT, and watching youtube.
- Define project goals, scope, and target audience.
  
Week 3-4
- Delayed due to other subjects, activities and projects that i need to complete.
- Use spare time to brainstorm ideas or refine goals, even if progress is limited.
  
Week-5 Preparation
- Start downloading the resources needed to build my application
- Set up a basic development environment.

Week-6 Design
- Create UI/UX designs

Week-7 Initial Development 
- Begin coding the application 

Week-8
- Still on the process of making my application

So, this is my first mobile app that i already proposed and this is also my first idea about making an app but, 
i didn't really put much thought about it at first, because it's already only a few left days before the deadline of 
proposing our idea or what mobile app should we decide to propose and I'm suddenly in a rush at that time of thinking 
and i don't have any idea of what mobile app should i propose. Then it leave me no choice but to proposing this app
because its really difficult for me to come up an idea about making a mobile app. After that in a week had passed, 
i also realized that this is going to be really difficult for me because i dont have a laptop, so i decided to change 
it, into something not really complicated and not really that difficult but also not easy mobile app.

# MB-R Generator
Now, this my new app that I'm going to make is a mobile app that generates recipes based on your mood that is why
its called MB-R Generator (Mood Based-Recipe Generator), it's an app that generates recipe based of how you feel like,
if you're happy it generates a recipe for your that could lightin up more of your mood.